+++ jac/jaclang/runtimelib/machine.py
@@ -5,16 +5,10 @@ from __future__ import annotations
 import inspect
 import marshal
 import os
-import copy
-import time
 import sys
 import tempfile
 import types
-import asyncio
-import dis
-import google.generativeai as genai
-import threading
-from collections import defaultdict
+
 from contextvars import ContextVar
 from typing import Optional, Union
 
@@ -29,6 +23,7 @@ from jaclang.runtimelib.architype import (
     NodeArchitype,
     WalkerArchitype,
 )
+from jaclang.runtimelib.gins.ghost import  ShellGhost
 from jaclang.utils.log import logging
 from jaclang.compiler.passes.main.schedules import py_code_gen, type_checker_sched
 
@@ -333,273 +328,3 @@ class JacProgram:
             return marshal.loads(result.ir.gen.py_bytecode)
         else:
             return None
-
-
-class CFGTracker:
-    def __init__(self):
-        self.executed_insts = {}
-        self.inst_lock = threading.Lock()
-        
-        self.curr_variables_lock = threading.Lock()
-        self.curr_variables = {}
-
-    def start_tracking(self):
-        """Start tracking branch coverage"""
-        frame = sys._getframe()
-        frame.f_trace_opcodes = True
-        sys.settrace(self.trace_callback)
-    def stop_tracking(self):
-        """Stop tracking branch coverage"""
-        sys.settrace(None)
-    
-    def get_exec_inst(self):
-        
-        self.inst_lock.acquire()
-        cpy = copy.deepcopy(self.executed_insts)
-        self.executed_insts = {}
-        self.inst_lock.release()
-        
-        return cpy
-
-    def get_variable_values(self):
-        self.curr_variables_lock.acquire()
-        cpy = copy.deepcopy(self.curr_variables)
-        self.curr_variables_lock.release()
-        
-        return cpy
-    
-    
-    def trace_callback(self, frame: types.FrameType, event: str, arg: any) -> Optional[types.TraceFunction]:
-        if event == "call":
-            frame.f_trace_opcodes = True
-        # frame.f_trace_lines = False
-    #     def tracefunc(frame, event, arg):
-    #         if event == 'call':
-    #             frame.f_trace_opcodes = True
-    #         elif event == 'opcode':
-    #             if frame.f_code.co_code[frame.f_lasti] == dis.opmap['MAKE_FUNCTION']:
-    #                 makefunctiontracefunc(frame)
-    # return tracefunc
-                
-        """Trace function to track executed branches"""        
-        code = frame.f_code
-        if ".jac" not in code.co_filename:
-            return self.trace_callback
-        
-
-
-        # if event != 'line':
-        #     return self.trace_callback
-        
-        if event == 'opcode':
-            # print(frame.f_lasti)
-        
-        # print(frame.f_lineno, frame.f_lasti)
-        
-        #edge case to handle executing code not within a function
-            filename = os.path.basename(code.co_filename)
-            module = code.co_name if code.co_name != "<module>" else os.path.splitext(filename)[0]
-    
-            self.inst_lock.acquire()
-            if module not in self.executed_insts:
-                self.executed_insts[module] = []
-            self.executed_insts[module].append(frame.f_lasti)
-            self.inst_lock.release()
-            variable_dict = {}
-            if '__annotations__' in frame.f_locals:
-                self.curr_variables_lock.acquire()
-                for var_name in frame.f_locals['__annotations__']:
-                    variable_dict[var_name] = frame.f_locals[var_name]
-                self.curr_variables[module] = (frame.f_lasti, variable_dict)
-                self.curr_variables_lock.release()
-
-        # self.variable_values[code.co_name][frame.f_lineno] = {}
-
-                
-        # print(f"{frame.f_lineno}")
-        # print(f"Function Name: {code.co_name}") 
-        # print(f"Filename: {code.co_filename}") 
-        # print(f"First Line Number: {code.co_firstlineno}") 
-        # print(f"Argument Count: {code.co_argcount}") 
-        # print(f"Constants: {code.co_consts}") 
-        # print(f"Local Variables: {code.co_varnames}")
-
-        # if event != 'line':
-        #     return self.trace_callback
-
-        # code = frame.f_code
-        # if code not in self.cfg_cache:
-        #     self.build_cfg(code)
-
-        # # Find current basic block
-        # blocks = self.cfg_cache[code]
-        # current_offset = frame.f_lasti
-
-        # # Find the block containing this offset
-        # current_block = None
-        # for block in blocks.values():
-        #     if block.offset <= current_offset <= block.offset + sum(inst.size for inst in block.instructions):
-        #         current_block = block
-        #         break
-
-        # if current_block:
-        #     current_block.hits += 1
-        #     # Record taken branches
-        #     for next_block in current_block.next:
-        #         self.coverage_data[code].add(
-        #             (current_block.offset, next_block.offset))
-
-        return self.trace_callback
-
-class ShellGhost:
-    def __init__(self):
-        self.cfgs = None
-        self.cfg_cv = threading.Condition()
-        self.tracker = CFGTracker()
-        self.sem_ir = None
-        
-        self.finished_exception_lock = threading.Lock()
-        self.exception = None
-        self.finished = False
-        self.variable_values = None
-
-        genai.configure(api_key=os.getenv("GEMINI_API_KEY"))
-        self.model = genai.GenerativeModel("gemini-1.5-flash")
-
-
-    def set_cfgs(self, cfgs):
-        self.cfg_cv.acquire()
-        self.cfgs = cfgs
-        self.cfg_cv.notify()
-        self.cfg_cv.release()
-
-    def start_ghost(self):
-        self.__ghost_thread = threading.Thread(target=self.worker)
-        self.__ghost_thread.start()
-
-    def set_finished(self, exception: Exception = None):
-        self.finished_exception_lock.acquire()
-        self.exception = exception
-        self.finished = True
-        self.finished_exception_lock.release()
-    
-    def prompt_llm(self, verbose: bool = False):
-        prompt = """I have a program.
-CFGS:
-{cfgs},
-Instructions per basic block:
-{instructions}
-Semantic and Type information from source code:
-{sem_ir}"""
-        
-        cfg_string = ""
-        ins_string = "" 
-        for module, cfg in self.cfgs.items():
-            cfg_string += f"Module: {module}\n{cfg}" 
-            ins_string += f"Module: {module}\n{cfg.display_instructions()}"
-        
-        prompt = prompt.format(cfgs=cfg_string, instructions=ins_string, sem_ir=self.sem_ir)
-
-        if self.variable_values != None:
-            prompt += "\nCurrent variable values at the specified bytecode offset:"
-            
-            for module, var_map in self.variable_values.items():
-                prompt += f"\nModule {module}: Offset: {var_map[0]}, Variables: {str(var_map[1])}"
-        
-        self.finished_exception_lock.acquire()
-        
-        if self.exception:
-            prompt += f"\nException: {self.exception}"
-
-        self.finished_exception_lock.release()
-        
-        prompt += "\nCan you identity bottlneck optimizations or where the code can error out?"
-        prompt += "\n(Reason about the program using bytecode, cfg, semantic and type information to infer the source code indirectly)"
-        
-        if verbose:
-            print(prompt)
-        
-        response = self.model.generate_content(prompt)
-        
-        print(response.text)
-    
-    def worker(self):
-
-        # get static cfgs
-        self.cfg_cv.acquire()
-        while (self.cfgs == None):
-            self.cfg_cv.wait()
-        # print(self.cfgs)
-        # for module_name, cfg in self.cfgs.items():
-        #     print(f"Name: {module_name}\n{cfg.display_instructions()}")
-        self.cfg_cv.release()
-        
-        # Once cv has been notifie, self.cfgs is no longer accessed across threads
-
-        
-        current_executing_bbs = {}   
-
-        def update_cfg():
-            exec_insts = self.tracker.get_exec_inst()
-            
-            # don't prompt if there's nothing new
-            if exec_insts == {}: return
-            for module, offset_list in exec_insts.items():
-                try: 
-                    cfg = self.cfgs[module]                    
-
-                    if module not in current_executing_bbs: # this means start at bb0, set exec count for bb0 to 1
-                        current_executing_bbs[module] = 0
-                        cfg.block_map.idx_to_block[0].exec_count = 1
-                    
-                    for offset in offset_list:                
-                        if offset not in cfg.block_map.idx_to_block[current_executing_bbs[module]].bytecode_offsets:
-                            for next in cfg.edges[current_executing_bbs[module]]:
-                                if offset in cfg.block_map.idx_to_block[next].bytecode_offsets:
-                                    cfg.edge_counts[(current_executing_bbs[module], next)] += 1
-                                    cfg.block_map.idx_to_block[next].exec_count += 1
-
-                                    current_executing_bbs[module] = next
-                                    break                                    
-                        assert(offset in cfg.block_map.idx_to_block[current_executing_bbs[module]].bytecode_offsets)
-                except Exception as e:
-                    self.set_finished(e)
-                    print(e)
-                    return
-
-            self.variable_values = self.tracker.get_variable_values()
-
-        self.finished_exception_lock.acquire()
-        while (not self.finished):
-            self.finished_exception_lock.release()
-            
-            time.sleep(5)
-            print('\nUpdating cfgs')
-            update_cfg()
-            self.prompt_llm()
-            self.finished_exception_lock.acquire()
-
-        self.finished_exception_lock.release()
-        
-        # print(self.cfgs)
-        # self.finished_exception_lock.acquire()
-        # if self.exception:
-            # print("Exception occured:", self.exception)
-        # self.finished_exception_lock.release()
-        print('\nUpdating cfgs at the end')
-        update_cfg()
-        self.prompt_llm()
-        
-        # response_dict = {'cfg': self.cfgs, 'instructions': self.cfgs[module_name].display_instructions(), 'list of local variables at sequential line numbers': variables_by_line}
-        # prompt = []
-        # for k,v in response_dict.items():
-        #     prompt.append(f"here is my {k}:\n{v}")
-        # prompt.append("\nCan you identify where the code could have an error?")
-        # response = model.generate_content("".join(prompt))
-
-        # print("PROMPT:\n")
-        # print("".join(prompt))
-        # print("RESPONSE:\n")
-        # print(response.text)
-
-        # print(self.cfgs['hot_path'].display_instructions())
